+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Fix force to add flags bug
 * [#157](https://github.com/tboox/xmake/issues/157): Fix generate pdb file error if it's output directory does not exists
+* Fix strip all symbols bug for macho target file
 
 ## v2.1.9
 
@@ -443,6 +444,7 @@
 
 * 修复无法通过`add_ldflags("xx", "xx", {force = true})`强制设置多个flags的问题
 * [#157](https://github.com/tboox/xmake/issues/157): 修复pdb符号输出目录不存在情况下编译失败问题
+* 修复对macho格式目标strip all符号失效问题
 
 ## v2.1.9